requirements.txt
discord.py
langchain
transformers